+++ src/index.tsx
@@ -29,6 +29,7 @@ export * from './layout/section/illustratedSection'
 export * from './layout/section/mediaSection'
 export * from './layout/section/slideSection'
 export * from './layout/section/textFieldsSection'
+export * from './layout/section/faqSection'
 
 export * from './layout/pages/seaPage'